fix(generate): move profile to separate file
